+++ src/components/themed/TransactionListTop.tsx
@@ -87,7 +87,7 @@ interface StateProps {
   dispatch: Dispatch
   displayDenomination: EdgeDenomination
   exchangeDenomination: EdgeDenomination
-  exchangeRate: string
+  exchangeRate: number
   exchangeRates: GuiExchangeRates
   isAccountBalanceVisible: boolean
   walletName: string
@@ -409,8 +409,8 @@ export class TransactionListTopComponent extends React.PureComponent<Props, Stat
 
     // Fiat Balance Formatting
     const exchangeAmount = convertNativeToDenomination(exchangeDenomination.multiplier)(nativeBalance)
-    const fiatBalance = mul(exchangeAmount, exchangeRate)
-    const fiatBalanceFormat = formatNumber(fiatBalance && gt(fiatBalance, '0.000001') ? fiatBalance : 0, { toFixed: 2 })
+    const fiatBalance = parseFloat(exchangeAmount) * exchangeRate
+    const fiatBalanceFormat = formatNumber(fiatBalance && fiatBalance > 0.000001 ? fiatBalance : 0, { toFixed: 2 })
 
     return (
       <>
@@ -783,7 +783,7 @@ export function TransactionListTop(props: OwnProps) {
       dispatch={dispatch}
       displayDenomination={displayDenomination}
       exchangeDenomination={exchangeDenomination}
-      exchangeRate={isKeysOnlyPlugin(wallet.currencyInfo.pluginId) ? '0' : exchangeRate}
+      exchangeRate={isKeysOnlyPlugin(wallet.currencyInfo.pluginId) ? 0 : exchangeRate}
       isAccountBalanceVisible={isAccountBalanceVisible}
       exchangeRates={exchangeRates}
       toggleBalanceVisibility={handleBalanceVisibility}